PerfectAI: Fix cpplint errors

+++ src/perfect/bufferedFile.h
@@ -21,18 +21,18 @@ private:
     HANDLE hFile; // Handle of the file
     unsigned int numThreads; // number of threads
     unsigned char* readBuffer; // Array of size [numThreads*blockSize] containing the data of the block, where reading is taking place
-    unsigned char* writeBuffer; //	 '' - access by [threadNo*bufferSize+position]
-    long long* curReadingPointer; // array of size [numThreads] with pointers to the byte which is currently read
-    long long* curWritingPointer; //			''
+    unsigned char* writeBuffer; // '' - access by [threadNo*bufferSize+position]
+    int64_t* curReadingPointer; // array of size [numThreads] with pointers to the byte which is currently read
+    int64_t* curWritingPointer; // ''
     unsigned int* bytesInReadBuffer; //
     unsigned int* bytesInWriteBuffer; //
     unsigned int bufferSize; // size in bytes of a buffer
-    long long fileSize; // size in bytes
+    int64_t fileSize; // size in bytes
     CRITICAL_SECTION csIO;
 
     // Functions
-    void writeDataToFile(HANDLE hFile, long long offset, unsigned int sizeInBytes, void* pData);
-    void readDataFromFile(HANDLE hFile, long long offset, unsigned int sizeInBytes, void* pData);
+    void writeDataToFile(HANDLE hFile, int64_t offset, unsigned int sizeInBytes, void* pData);
+    void readDataFromFile(HANDLE hFile, int64_t offset, unsigned int sizeInBytes, void* pData);
 
 public:
     // Constructor / destructor
@@ -43,9 +43,9 @@ public:
     bool flushBuffers();
     bool writeBytes(unsigned int numBytes, unsigned char* pData);
     bool readBytes(unsigned int numBytes, unsigned char* pData);
-    bool writeBytes(unsigned int threadNo, long long positionInFile, unsigned int numBytes, unsigned char* pData);
-    bool readBytes(unsigned int threadNo, long long positionInFile, unsigned int numBytes, unsigned char* pData);
-    long long getFileSize();
+    bool writeBytes(unsigned int threadNo, int64_t positionInFile, unsigned int numBytes, unsigned char* pData);
+    bool readBytes(unsigned int threadNo, int64_t positionInFile, unsigned int numBytes, unsigned char* pData);
+    int64_t getFileSize();
 };
 
 #endif // BUFFERED_FILE_H_INCLUDED
