+++ src/perfect/miniMax_retroAnalysis.h
@@ -20,7 +20,7 @@ struct RetroAnalysisThreadVars // thread specific variables for each thread in t
 {
     vector<CyclicArray*> statesToProcess; // vector-queue containing the states, whose short knot value are known for sure. they have to be processed. if processed the state will be removed from list. indexing: [threadNo][plyNumber]
     vector<vector<RetroAnalysisQueueState>> stateQueue; // Queue containing states, whose 'count value' shall be increased by one. Before writing 'count value' to 'count array' the writing positions are sorted for faster processing.
-    long long numStatesToProcess; // Number of states in 'statesToProcess' which have to be processed
+    int64_t numStatesToProcess; // Number of states in 'statesToProcess' which have to be processed
     unsigned int threadNo;
 };
 
@@ -30,8 +30,8 @@ struct RetroAnalysisVars // constant during calculation
     vector<compressorClass::compressedArrayClass*> countArraysCompr; // '' but compressed
     vector<bool> layerInitialized; //
     vector<unsigned int> layersToCalculate; // layers which shall be calculated
-    long long totalNumKnots; // total numbers of knots which have to be stored in memory
-    long long numKnotsToCalc; // number of knots of all layers to be calculated
+    int64_t totalNumKnots; // total numbers of knots which have to be stored in memory
+    int64_t numKnotsToCalc; // number of knots of all layers to be calculated
     vector<RetroAnalysisThreadVars> thread;
 };
